Tweak solargraph
Solargraph is now set to use a named install of Ruby rather than
deferring to bundler. The alternative to this is either manually
installing solargraph on each new version or polluting every project
`Gemfile` with the gems, which aren't actually part of the project.

Enable formatting, but disable format-on-save so I can quickly apply
style updates via `call CocAction('format')`