Move mapleader to the top of the config
Mapleader isn't a dynamic value, when `<leader>` is given in a mapping
it's substituted with the `mapleader` value at that moment. I need to
move my mapleader from the old config to the new config so any new
`<leader>` mappings get defined with the leader `<space>`.

Add `<leader>t` to call coc's file formatter method. I don't want to
reformat files on save because it'll absolutely ruin any commits on
legacy projects where I'm changing 5 words in a 300 line class.